+++ src/allpress/net/scrape.py
@@ -109,6 +109,7 @@ class Scraper:
         :param iterations: The number of recursive iterations to perform.
         :return:
         """
+
         iteration = 0
         self.starting_url = domain
         start_page = requests.get(domain)
@@ -117,6 +118,7 @@ class Scraper:
         to_scrape = list(initial_links)
 
         while iteration < iterations:
+            new_soups = {}
             iteration += 1
             logger.info(f'Scraping iteration {iteration}...')
 
@@ -139,6 +141,8 @@ class Scraper:
 
                     soup = Soup(response.content, 'html.parser')
                     self.scraped_urls.add(url)
+                    title = soup.title.string if soup.title else 'No Title'
+                    new_soups[title] = (soup, url)
 
                     # Caches the scraped URL if it's an article.
                     # The confidence threshold of the ArticleDetector can be adjusted.
@@ -159,6 +163,7 @@ class Scraper:
                     continue
 
             to_scrape = list(new_found_urls - self.scraped_urls)
+            yield new_soups
         logger.info(f'Iteration {iteration} done. Scraping {len(to_scrape)} URLs in Iteration {iteration+1}, scraped {len(self.scraped_urls)} URLs.')