+++ src/allpress/settings.py
@@ -1,0 +1,7 @@
+
+CLASSIFICATION_MODELS_PATH = "C:\\Users\\preit\\OneDrive\\Desktop\\coding projects\\Allpress2\\src\\allpress\\models"
+
+DATABASE_USERNAME = "test"
+DATABASE_PASSWORD = "default"
+DATABASE_HOST = "localhost"
+DATABASE_NAME = "test"