+++ src/changelog.md
@@ -1,6 +1,14 @@
 # July 2025
 
-### 0.1.1a
+### 0.2.0a (2025-07-01)
+
+- Add `Article` object to represent articles
+- Add AutoEncoder for semantic and rhetorical representation.
+
+### 0.1.1a (2025-07-01)
+
+- Add documentation for `scrape()` method of `Scraper`
+- Small code changes
 
 ### 0.1.0a (2025-07-01)